+++ libs/fsm/tests/redispatch.c
@@ -95,9 +95,9 @@ static enum am_rc redisp_fsm_sinit(
 
 static void redispatch_fsm(void) {
     struct redisp_fsm* me = &m_redisp_fsm;
-    am_fsm_create(&me->fsm, redisp_fsm_sinit);
+    am_fsm_init(&me->fsm, redisp_fsm_sinit);
 
-    am_fsm_init(&me->fsm, /*init_event=*/NULL);
+    am_fsm_start(&me->fsm, /*init_event=*/NULL);
     AM_ASSERT(0 == me->a_handled);
     AM_ASSERT(0 == me->b_handled);